Move suiteapp_file_cabinet into client folder
