+++ packages/netsuite-adapter/test/adapter.test.ts
@@ -34,9 +34,9 @@ import { CustomizationInfo, CustomTypeInfo, FileCustomizationInfo, FolderCustomi
 import * as changesDetector from '../src/changes_detector/changes_detector'
 import SuiteAppClient from '../src/client/suiteapp_client/suiteapp_client'
 import { SERVER_TIME_TYPE_NAME } from '../src/server_time'
-import * as suiteAppFileCabinet from '../src/suiteapp_file_cabinet'
+import * as suiteAppFileCabinet from '../src/client/suiteapp_client/suiteapp_file_cabinet'
 import { SDF_CREATE_OR_UPDATE_GROUP_ID } from '../src/group_changes'
-import { SuiteAppFileCabinetOperations } from '../src/suiteapp_file_cabinet'
+import { SuiteAppFileCabinetOperations } from '../src/client/suiteapp_client/suiteapp_file_cabinet'
 import getChangeValidator from '../src/change_validator'
 import { FetchByQueryFunc } from '../src/change_validators/safe_deploy'
 import { getStandardTypesNames } from '../src/autogen/types'
