+++ packages/netsuite-adapter/test/client/suiteapp_client/suiteapp_file_cabinet.test.ts
@@ -17,13 +17,13 @@ import { Change, InstanceElement, StaticFile, toChange } from '@salto-io/adapter
 import _ from 'lodash'
 import { MockInterface } from '@salto-io/test-utils'
 import { collections } from '@salto-io/lowerdash'
-import { NetsuiteQuery } from '../src/query'
-import SuiteAppClient from '../src/client/suiteapp_client/suiteapp_client'
-import { THROW_ON_MISSING_FEATURE_ERROR, createSuiteAppFileCabinetOperations, isChangeDeployable } from '../src/suiteapp_file_cabinet'
-import { ReadFileEncodingError, ReadFileError, ReadFileInsufficientPermissionError } from '../src/client/suiteapp_client/errors'
-import { customtransactiontypeType } from '../src/autogen/types/standard_types/customtransactiontype'
-import { ExistingFileCabinetInstanceDetails, FileCabinetInstanceDetails } from '../src/client/suiteapp_client/types'
-import { getFileCabinetTypes } from '../src/types/file_cabinet_types'
+import { NetsuiteQuery } from '../../../src/query'
+import SuiteAppClient from '../../../src/client/suiteapp_client/suiteapp_client'
+import { createSuiteAppFileCabinetOperations, isChangeDeployable } from '../../../src/client/suiteapp_client/suiteapp_file_cabinet'
+import { ReadFileEncodingError, ReadFileError, ReadFileInsufficientPermissionError } from '../../../src/client/suiteapp_client/errors'
+import { customtransactiontypeType } from '../../../src/autogen/types/standard_types/customtransactiontype'
+import { ExistingFileCabinetInstanceDetails, FileCabinetInstanceDetails } from '../../../src/client/suiteapp_client/types'
+import { getFileCabinetTypes } from '../../../src/types/file_cabinet_types'
 
 const { awu } = collections.asynciterable
 
@@ -358,24 +358,15 @@ describe('suiteapp_file_cabinet', () => {
     })
 
     it('should filter files with query', async () => {
-      query.isFileMatch.mockImplementation(path => path !== '/folder5/folder3/file1')
+      query.isFileMatch.mockImplementation(path => path !== '/folder5/folder4' && path !== '/folder5/folder3/file1')
       const { elements } = await createSuiteAppFileCabinetOperations(suiteAppClient)
         .importFileCabinet(query)
       expect(elements).toEqual([
         expectedResults[0],
         expectedResults[1],
-        expectedResults[2],
         expectedResults[4],
         expectedResults[5],
       ])
-    })
-
-    it('should call suiteql with missing feature error param', async () => {
-      await createSuiteAppFileCabinetOperations(suiteAppClient).importFileCabinet(query)
-      expect(mockSuiteAppClient.runSuiteQL).toHaveBeenCalledWith(
-        expect.stringContaining('SELECT name, id, bundleable'),
-        THROW_ON_MISSING_FEATURE_ERROR
-      )
     })
 
     it('return empty result when no top level folder matches the adapter\'s query', async () => {
@@ -473,22 +464,6 @@ describe('suiteapp_file_cabinet', () => {
 
       await expect(createSuiteAppFileCabinetOperations(suiteAppClient)
         .importFileCabinet(query)).rejects.toThrow()
-    })
-
-    it('should remove excluded folder before creating the file cabinet query', async () => {
-      query.isFileMatch.mockImplementation(path => !path.includes('folder4'))
-      const { elements } = await createSuiteAppFileCabinetOperations(suiteAppClient)
-        .importFileCabinet(query)
-      const testWhereQuery = 'hideinbundle = \'F\' AND folder IN (5, 3)'
-      const suiteQlQuery = 'SELECT name, id, filesize, bundleable, isinactive, isonline,'
-      + ' addtimestamptourl, hideinbundle, description, folder, islink, url'
-      + ` FROM file WHERE ${testWhereQuery} ORDER BY id ASC`
-      expect(suiteAppClient.runSuiteQL).toHaveBeenNthCalledWith(3, suiteQlQuery)
-      expect(elements).toEqual([
-        expectedResults[0],
-        expectedResults[1],
-        expectedResults[3],
-      ])
     })
   })
